requirements.txt
black[jupyter]==22.3.0
cupy_cuda113==10.6.0
einops==0.6.1
future==0.18.3
imageio==2.31.2
ipykernel==6.16.2
ipywidgets==8.1.1
jupyter==1.0.0
kornia==0.6.12
lpips==0.1.4
matplotlib==3.5.3
numpy==1.20.0
opencv_contrib_python==4.7.0.72
pillow==9.5.0
pipreqs==0.4.13
plotly==5.18.0
pycocotools==2.0.6
pytorch_msssim==0.2.1
scikit_learn==1.0.2
scikit_image==0.18.3
scipy==1.7.3
setuptools==59.5.0  # fix error, see https://github.com/pytorch/pytorch/issues/69894
shapely==2.0.2
tensorboard
tqdm==4.61.2
trimesh==3.21.*
pandas==1.3.5
xgboost==1.6.2
dtreeviz==2.2.1
wand==0.6.11
pyiqa==0.1.6.4
isort==5.11.5
